(skip rds db on aarch64)

+++ buildspecs/delete-test-db.sh
@@ -1,5 +1,11 @@
 #!/bin/sh
 
+if [ $ARCH == "aarch64" ]; then
+    # https://docs.cypress.io/guides/getting-started/installing-cypress#Download-URLs
+    echo "Skipping this step on aarch64 because we're not running Cypress tests there."
+    exit 0
+fi
+
 aws rds delete-db-instance \
     --db-instance-identifier astrobin-test-${CODEBUILD_BUILD_NUMBER}
     --skip-final-snapshot \
